requirements.txt
Pillow== 10.1.0